ImportantIndicators/pyproject.toml
[project]
name = "important-indicators"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=1.0",
    "curl-cffi",
    "fredapi",
    "python-dotenv",
    "postgrest",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
